Aplicando conexion de la base de datos e instalacion de paquetes NuGets

+++ Sis457Restaurante/CadRestaurante/Categoria.cs
@@ -24,7 +24,7 @@ namespace CadRestaurante
         public string nombre { get; set; }
         public string usuarioRegistro { get; set; }
         public System.DateTime fechaRegistro { get; set; }
-        public int estado { get; set; }
+        public short estado { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
         public virtual ICollection<Producto> Producto { get; set; }
