+++ Sis457Restaurante/CadRestaurante/Cliente.cs
@@ -25,7 +25,7 @@ namespace CadRestaurante
         public string razonSocial { get; set; }
         public string usuarioRegistro { get; set; }
         public System.DateTime fechaRegistro { get; set; }
-        public int estado { get; set; }
+        public short estado { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
         public virtual ICollection<Venta> Venta { get; set; }
