+++ Sis457Restaurante/CadRestaurante/Usuario.cs
@@ -20,7 +20,7 @@ namespace CadRestaurante
         public string clave { get; set; }
         public string usuarioRegistro { get; set; }
         public System.DateTime fechaRegistro { get; set; }
-        public int estado { get; set; }
+        public short estado { get; set; }
     
         public virtual Empleado Empleado { get; set; }
     }
